5251 - Splitting classes out and added Crown Copyright
